+++ business/src/main/java/com/kevinguanchedarias/owgejava/business/SpeedImpactGroupBo.java
@@ -1,6 +1,6 @@
 package com.kevinguanchedarias.owgejava.business;
 
-import com.kevinguanchedarias.owgejava.business.unit.util.UnitTypeInheritanceFinderService;
+import com.kevinguanchedarias.owgejava.business.speedimpactgroup.SpeedImpactGroupFinderBo;
 import com.kevinguanchedarias.owgejava.dto.ObjectRelationDto;
 import com.kevinguanchedarias.owgejava.dto.SpeedImpactGroupDto;
 import com.kevinguanchedarias.owgejava.entity.EntityWithMissionLimitation;
@@ -9,15 +9,14 @@ import com.kevinguanchedarias.owgejava.entity.ObjectRelationToObjectRelation;
 import com.kevinguanchedarias.owgejava.entity.RequirementGroup;
 import com.kevinguanchedarias.owgejava.entity.SpeedImpactGroup;
 import com.kevinguanchedarias.owgejava.entity.Unit;
-import com.kevinguanchedarias.owgejava.entity.UnitType;
 import com.kevinguanchedarias.owgejava.entity.UserStorage;
 import com.kevinguanchedarias.owgejava.enumerations.ObjectEnum;
 import com.kevinguanchedarias.owgejava.repository.ObjectRelationToObjectRelationRepository;
 import com.kevinguanchedarias.owgejava.repository.SpeedImpactGroupRepository;
 import com.kevinguanchedarias.owgejava.util.DtoUtilService;
 import com.kevinguanchedarias.taggablecache.manager.TaggableCacheManager;
-import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -63,7 +62,8 @@ public class SpeedImpactGroupBo implements BaseBo<Integer, SpeedImpactGroup, Spe
     private transient TaggableCacheManager taggableCacheManager;
 
     @Autowired
-    private transient UnitTypeInheritanceFinderService unitTypeInheritanceFinderService;
+    @Lazy
+    private transient SpeedImpactGroupFinderBo speedImpactGroupFinderBo;
 
     @Override
     public JpaRepository<SpeedImpactGroup, Integer> getRepository() {
@@ -149,13 +149,8 @@ public class SpeedImpactGroupBo implements BaseBo<Integer, SpeedImpactGroup, Spe
         return speedImpactGroups.stream().map(EntityWithMissionLimitation::getId).collect(Collectors.toList());
     }
 
-    public boolean canIntercept(List<InterceptableSpeedGroup> interceptableSpeedGroups, Unit unit) {
-        SpeedImpactGroup speedImpactGroup = ObjectUtils.getFirstNonNull(
-                unit::getSpeedImpactGroup,
-                () -> unitTypeInheritanceFinderService.findUnitTypeMatchingCondition(
-                        unit.getType(), unitType -> unitType.getSpeedImpactGroup() != null
-                ).orElse(new UnitType()).getSpeedImpactGroup()
-        );
+    public boolean canIntercept(List<InterceptableSpeedGroup> interceptableSpeedGroups, UserStorage user, Unit unit) {
+        SpeedImpactGroup speedImpactGroup = speedImpactGroupFinderBo.findApplicable(user, unit);
         return speedImpactGroup != null && interceptableSpeedGroups.stream().anyMatch(current -> current.getSpeedImpactGroup().getId()
                 .equals(speedImpactGroup.getId()));
     }